requirement.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
jinja2
groq
